requirements.txt
streamlit
opencv-python-headless
pillow
pdf2image
pytesseract
pandas
numpy
